requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django==1.4.3
South==0.7.6


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
Fabric
django-libs
factory-boy
mock
django-nose
coverage
django-coverage
ipython
ipdb
